some fixes in circular .h files

+++ module_9_forest/module_9_forest/Forest.cpp
@@ -11,7 +11,8 @@ Forest::Forest()
 
 void Forest::growUp(const Tree& nameTree)
 {
-	std::cout << "A tree grew in the forest : " << nameTree.getNameTree() << std::endl;
+	std::cout << "A tree grew in the forest : " 
+		<< static_cast<int>(nameTree.getNameTree()) << std::endl;
 
 }
